refactor: separate CustomLegend component
