+++ src/components/PopulationChart.tsx
@@ -13,6 +13,7 @@ import {
 import { usePopulationData } from '@/hooks/usePopulationData';
 import { useRef, useEffect, useState } from 'react';
 import { Prefecture } from '@/types/interfaces';
+import CustomLegend from './CustomLegend';
 
 export default function PopulationChart({
   prefectures,
@@ -35,117 +36,24 @@ export default function PopulationChart({
   const populationData = data?.populationData || [];
   const boundaryYears = data?.boundaryYears || {};
 
-  // Define our custom legend item type.
-  interface CustomLegendItem {
-    value?: string | number;
-    color?: string;
-    dataKey?: string | number;
-    payload?: {
-      legendType?: string;
-    };
-  }
-
-  // Define the props that our custom legend expects.
-  interface CustomLegendProps {
-    payload?: CustomLegendItem[];
-  }
-
-  // Our custom legend component.
-  const CustomLegend = (props: CustomLegendProps) => {
-    const { payload = [] } = props;
-    const containerRef = useRef<HTMLDivElement>(null);
-    const [scrollingDown, setScrollingDown] = useState(true);
-
-    useEffect(() => {
-      const scrollContainer = containerRef.current;
-      if (!scrollContainer) return;
-
-      if (selectedPrefectures.length < 8) return;
-
-      const scrollStep = 1;
-      const intervalTime = 58 - selectedPrefectures.length;
-      let interval: ReturnType<typeof setInterval>;
-
-      const startScrolling = () => {
-        interval = setInterval(() => {
-          if (scrollingDown) {
-            scrollContainer.scrollTop += scrollStep;
-            if (
-              scrollContainer.scrollTop + scrollContainer.clientHeight >=
-              scrollContainer.scrollHeight
-            ) {
-              setScrollingDown(false);
-            }
-          } else {
-            scrollContainer.scrollTop -= scrollStep;
-            if (scrollContainer.scrollTop <= 0) {
-              setScrollingDown(true);
-            }
-          }
-        }, intervalTime);
-      };
-
-      startScrolling();
-
-      const stopAutoScroll = () => {
-        clearInterval(interval);
-      };
-
-      scrollContainer.addEventListener('mouseenter', stopAutoScroll);
-      scrollContainer.addEventListener('mouseleave', startScrolling);
-
-      return () => {
-        scrollContainer.removeEventListener('mouseenter', stopAutoScroll);
-        scrollContainer.removeEventListener('mouseleave', startScrolling);
-        stopAutoScroll();
-      };
-    }, [scrollingDown]);
-
-    return (
-      <div
-        className="hide-scrollbar flex h-[500px] flex-wrap justify-between space-y-2 overflow-y-scroll py-50 pr-2 pl-2 lg:pr-4"
-        ref={containerRef}
-      >
-        <div className="pointer-events-none absolute top-0 left-0 h-40 w-full bg-gradient-to-b from-white via-white/90 to-transparent" />
-        <div className="pointer-events-none absolute top-0 right-0 h-full w-0.5 bg-gradient-to-b from-white via-gray-300 to-white" />
-
-        {payload
-          // Check that payload exists before accessing legendType.
-          .filter((entry: CustomLegendItem) =>
-            entry.payload ? entry.payload.legendType !== 'none' : false
-          )
-          .map((entry: CustomLegendItem) => (
-            <div
-              key={`item-${entry.value}`}
-              className="flex min-w-25 items-center text-lg"
-            >
-              <span
-                className="mr-0.5 inline-block h-0.5 w-3"
-                style={{ backgroundColor: entry.color }}
-              ></span>
-              <span
-                className="mr-0.5 inline-block h-0.5 w-0.5"
-                style={{ backgroundColor: entry.color }}
-              ></span>
-              <span
-                className="mr-2 inline-block h-0.5 w-0.5"
-                style={{ backgroundColor: entry.color }}
-              ></span>
-              {entry.value !== undefined ? getPrefectureName(entry.value) : ''}
-            </div>
-          ))}
-        <div className="pointer-events-none absolute bottom-0 left-0 h-40 w-full bg-linear-to-t from-white via-white/90 to-transparent" />
-      </div>
-    );
+  // getPrefectureName now accepts a non-undefined key.
+  const getPrefectureName = (key: string | number): string => {
+    if (typeof key === 'number') {
+      const pref = prefectures.find((p) => p.prefCode === key);
+      return pref ? pref.prefName : String(key);
+    }
+    // Assume the key is in the format 'Pref-<code>'
+    const code = parseInt(key.replace('Pref-', ''), 10);
+    const pref = prefectures.find((p) => p.prefCode === code);
+    return pref ? pref.prefName : key;
   };
-  const legendHeight = 500;
 
   // Ensure tick formatter always returns a string.
   const formatYAxis = (tick: number): string => {
     return tick > 10000 ? (tick / 10000).toString() + '万' : tick.toString();
   };
 
-  // Define the style type for the legend wrapper, using a literal union for textAlign.
+  // Define the style type for the legend wrapper.
   interface LegendWrapperStyle {
     position: 'absolute';
     maxWidth: string;
@@ -156,6 +64,7 @@ export default function PopulationChart({
     textAlign: 'left' | 'center' | 'right';
   }
 
+  const legendHeight = 500;
   const legendWrapperStyle: LegendWrapperStyle = {
     position: 'absolute',
     maxWidth: '120px',
@@ -168,22 +77,15 @@ export default function PopulationChart({
 
   const getLegendProps = () => {
     return {
-      content: <CustomLegend />,
+      content: (
+        <CustomLegend
+          selectedPrefectures={selectedPrefectures}
+          getPrefectureName={getPrefectureName}
+        />
+      ),
       height: 1,
       wrapperStyle: legendWrapperStyle,
     };
-  };
-
-  // getPrefectureName now accepts a non-undefined key.
-  const getPrefectureName = (key: string | number): string => {
-    if (typeof key === 'number') {
-      const pref = prefectures.find((p) => p.prefCode === key);
-      return pref ? pref.prefName : String(key);
-    }
-    // Assume the key is in the format 'Pref-<code>'
-    const code = parseInt(key.replace('Pref-', ''), 10);
-    const pref = prefectures.find((p) => p.prefCode === code);
-    return pref ? pref.prefName : key;
   };
 
   const colorPalette = [
